refactor(engine): split into S2AutomateEngine + S2AutomateOutcomeEngine + Phase B fixes
The 418-line S2AutomateEngineImpl conflated two contracts (fail-fast
vs per-item-outcomes). Split into:
- S2AutomateEngineBase (abstract; shared infra)
- S2AutomateEngineImpl (legacy, ~96 LOC)
- S2AutomateOutcomeEngineImpl (outcomes; baked Phase B fixes)

Phase B reactivity fixes in the new outcome engine:
- B.1 createWithOutcomes uses chunked flow instead of unbounded toList
- B.2 doTransitionWithOutcomes uses batched load(ids) per chunk
  (~500ms->~5ms latency for 100-cmd Mongo batch)
- B.3 outcome->context correlation by commandId (O(N2)->O(N))

S2ConfigurerAdapter absorbs dual-engine wiring so downstream
extension points (c2's S2SsmConfigurerAdapter, ssm's
DeliveryStoringAutomateConfig, samples) need zero source change.
